requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
passlib
python-jose[cryptography]
python-multipart